Time: 35 ms (100.00%), Space: 6.5 MB (100.00%) - LeetHub

+++ 546-remove-boxes/546-remove-boxes.rs
@@ -10,21 +10,21 @@ impl Solution {
     pub fn remove_boxes(boxes: Vec<i32>) -> i32 {
         let n = boxes.len();
         let mut dp = vec![vec![vec![0; n]; n]; n];
-        remove_boxes(&mut dp, &boxes, 0, n - 1, 0) as i32
+        remove_boxes(&mut dp, &boxes, 0, n - 1, 0)
     }
 }
 
 fn remove_boxes(
-    dp: &mut Vec<Vec<Vec<usize>>>,
+    dp: &mut Vec<Vec<Vec<i32>>>,
     boxes: &Vec<i32>,
     i: usize,
     j: usize,
     k: usize,
-) -> usize {
+) -> i32 {
     if i > j {
         return 0;
     } else if i == j {
-        return (k + 1) * (k + 1);
+        return ((k + 1) * (k + 1)) as i32;
     } else if dp[i][j][k] > 0 {
         return dp[i][j][k];
     }
